Worked on Front-end
I modified on the task-6-prototype.md file, which covers the commands I used to install dependencies, and added the commands I used to make the front-end. I also started to work on a basic front-end code which will end up displaying data from a database, which will be important for the overall project my team is working on.