Working on Task #8 "Inventory Data Requirements"
For task https://tree.taiga.io/project/ctbocchi-street-cards-capstone-project/task/8

To do this task, I decided to use this - the prototype I made for Task #6, to complete this task, doing proper checks and all, as it is a part of the "Research" progress board and not on a deliverable progress board, which is for the main project.

I plan on later on adding checks for each of the pieces of information to ensure the right data is entered for each field when someone uses Django REST Framework.