(Task #8) Implemented checks for dates.
The last thing I did was rather painful to figure out, but it involved ensuring that the user (staff) cannot submit dates for 'date_checked_out' or 'check_out_by' earlier than 'date_checked_in', which can be a major error. I could also make it where 'check_out_by' is not earlier than 'date_checked_out', but it makes more sense for that to happen, and the way of resolving that issue by other means than checking POST requests.